+++ src/SkillSearchMainForm.pas
@@ -9,6 +9,7 @@ type
   TSkillSearchMainForm = class(TForm)
   private
     fInfoLabel: TLabel;
+    procedure InitializeDatabase;
   public
     constructor Create(aOwner: TComponent); override;
   end;
@@ -18,6 +19,9 @@ var
 
 implementation
 
+uses
+  AppPaths, DatabaseManager;
+
 constructor TSkillSearchMainForm.Create(aOwner: TComponent);
 begin
   inherited Create(aOwner);
@@ -25,11 +29,25 @@ begin
   Width := 960;
   Height := 640;
 
+  InitializeDatabase;
+
   fInfoLabel := TLabel.Create(self);
   fInfoLabel.Parent := self;
   fInfoLabel.Left := 24;
   fInfoLabel.Top := 24;
   fInfoLabel.Caption := 'Project bootstrap complete. Core features are implemented in subsequent tasks.';
+end;
+
+procedure TSkillSearchMainForm.InitializeDatabase;
+var
+  lDbManager: TDatabaseManager;
+begin
+  lDbManager := TDatabaseManager.Create(GetCacheDbPath, GetSqliteDllPath);
+  try
+    lDbManager.Initialize;
+  finally
+    lDbManager.Free;
+  end;
 end;
 
 end.